requirements.txt
fastapi
uvicorn
openai
langchain
pinecone-client
python-dotenv